+++ src/main/java/refinedstorage/RefinedStorage.java
@@ -139,4 +139,8 @@ public final class RefinedStorage {
     public static boolean hasIC2() {
         return Loader.isModLoaded("IC2");
     }
+
+    public static boolean hasTesla() {
+        return Loader.isModLoaded("Tesla");
+    }
 }